project expense-tracker/requirement.txt
streamlit== 1.45.0
fastapi == 0.115.12
fastapi-cli == 0.0.7
pydantic == 2.11.1
uvicorn == 0.34.0
mysql-connector-python == 9.3.0
requests ==2.32.3
pandas== 2.2.3
pytest== 8.3.5